automatizacion_reportes/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
polars
python-docx
